move show schema result to schema module
