+++ server/src/main/java/org/apache/iotdb/db/metadata/plan/schemaregion/result/ShowTimeSeriesResult.java
@@ -1,0 +1,114 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.iotdb.db.metadata.plan.schemaregion.result;
+
+import org.apache.iotdb.tsfile.file.metadata.enums.CompressionType;
+import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
+import org.apache.iotdb.tsfile.file.metadata.enums.TSEncoding;
+
+import java.util.Map;
+import java.util.Objects;
+
+public class ShowTimeSeriesResult extends ShowSchemaResult {
+
+  private String alias;
+  private TSDataType dataType;
+  private TSEncoding encoding;
+  private CompressionType compressor;
+  private Map<String, String> tags;
+  private Map<String, String> attributes;
+
+  private String deadband;
+  private String deadbandParameters;
+
+  public ShowTimeSeriesResult(
+      String name,
+      String alias,
+      String sgName,
+      TSDataType dataType,
+      TSEncoding encoding,
+      CompressionType compressor,
+      Map<String, String> tags,
+      Map<String, String> attributes,
+      String deadband,
+      String deadbandParameters) {
+    super(name, sgName);
+    this.alias = alias;
+    this.dataType = dataType;
+    this.encoding = encoding;
+    this.compressor = compressor;
+    this.tags = tags;
+    this.attributes = attributes;
+    this.deadband = deadband;
+    this.deadbandParameters = deadbandParameters;
+  }
+
+  public ShowTimeSeriesResult() {
+    super();
+  }
+
+  public String getAlias() {
+    return alias;
+  }
+
+  public TSDataType getDataType() {
+    return dataType;
+  }
+
+  public TSEncoding getEncoding() {
+    return encoding;
+  }
+
+  public CompressionType getCompressor() {
+    return compressor;
+  }
+
+  public Map<String, String> getTag() {
+    return tags;
+  }
+
+  public Map<String, String> getAttribute() {
+    return attributes;
+  }
+
+  public String getDeadband() {
+    return deadband;
+  }
+
+  public String getDeadbandParameters() {
+    return deadbandParameters;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    ShowTimeSeriesResult result = (ShowTimeSeriesResult) o;
+    return Objects.equals(path, result.path);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(path);
+  }
+}
